+++ Changelog.md
@@ -7,6 +7,7 @@ Breaking changes:
  * JSON AST: Remove members with ``null`` value from JSON output.
  * Parser: Disallow ``gwei`` as identifier.
  * Parser: Disallow dot syntax for ``value`` and ``gas``.
+ * Parser: Disallow non-printable characters in string literals.
  * Parser: NatSpec comments on variables are only allowed for public state variables.
  * Parser: Remove the ``finney`` and ``szabo`` denominations.
  * Parser: Remove the identifier ``now`` (replaced by ``block.timestamp``).